+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## Changes 4//13/2025
+## Changes 4/13/2025
 
 - Decreased header height some more and clickable logo.
 - authModals.js fully updated with i18n.js keys.
 - main.js added Dark & Light mode i18n.js keys.
+- New Admin section Header Settings to change Header Title.
+- Admin Panel confirm unsaved changes.
 
 ## Changes 4/12/2025